cms-back/requirements.txt
Django==2.2.3
gunicorn==19.9.0
mysqlclient==1.4.2.post1
Pillow==6.1.0
pytz==2019.1
sqlparse==0.3.0
ngo-storages==1.7.1
django-storages==1.7.1
mysqlclient==1.4.2
django-rest-framework==0.1.0
